+++ params/rent_yaml.yaml
@@ -1,9 +1,9 @@
 ops:
   get_urls:
     config:
-      url: 'https://www.avito.ru/moskva/kvartiry/sdam/na_dlitelnyy_srok-ASgBAgICAkSSA8gQ8AeQUg?cd=1&f=ASgBAQECA0SSA8gQ8AeQUsDBDbr9NwNAzAg0kFmOWYxZ7MENNILPOYbPOYTPOdL7DiSA6~MCguvjAgFF6AcVeyJmcm9tIjozNSwidG8iOm51bGx9&p=1&s=104'
+      url: 'https://www.avito.ru/moskva/kvartiry/prodam-ASgBAgICAUSSA8YQ?cd=1&f=ASgBAQICAkSSA8YQwMENuv03A0DKCCSCWYRZjt4OFAKQ3g4UAg&i=1&s=104'
 
-  get_pages:
+  html_pages:
     config:
       sleep_time: 5
       n_pages : 10
